added shared test helpers to new file
